# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/dev-token": {
      "post": {
        "operationId": "AuthController_devToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Jeton de développement (mode dev uniquement)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Connexion d'un compte géré (matricule + code/mot de passe)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profil du compte connecté (nom, grade, rôles, photo)",
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifier son profil : nom affiché et/ou photo (audité)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/select-role": {
      "post": {
        "operationId": "AuthController_selectRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Choisir le rôle actif (compte multi-rôles) — nouveau jeton",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changer son mot de passe (1er login) — active le compte",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "ARGOS API",
    "description": "API du poste de commandement ARGOS — Phase 0/1 : IAM/RBAC (default-deny), audit chaîné, feature flags.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "DevTokenDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "k.benjelloun"
          },
          "role": {
            "type": "string",
            "enum": [
              "superadmin",
              "admin",
              "strategic",
              "tacom",
              "bluecell",
              "greencell",
              "orangecell",
              "resp_hospital",
              "resp_shelter",
              "resp_morgue",
              "resp_unit",
              "resp_equipment"
            ],
            "example": "superadmin"
          }
        },
        "required": [
          "username",
          "role"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "matricule": {
            "type": "string",
            "example": "n.fassi"
          },
          "password": {
            "type": "string",
            "example": "A7X2-K9D3"
          }
        },
        "required": [
          "matricule",
          "password"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "description": "Nom affiché"
          },
          "photo": {
            "type": "string",
            "nullable": true,
            "description": "Photo de profil (data URL) ; null pour retirer"
          }
        }
      },
      "SelectRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "superadmin",
              "admin",
              "strategic",
              "tacom",
              "bluecell",
              "greencell",
              "orangecell",
              "resp_hospital",
              "resp_shelter",
              "resp_morgue",
              "resp_unit",
              "resp_equipment"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
    }
  }
}